Int instead of long, some more logging

+++ 09-timeouts/src/finance/Events/InvoiceIssued.cs
@@ -4,7 +4,7 @@ namespace Finance.Events
 {
     public class InvoiceIssued
     {
-        public long InvoiceNumber { get; set; }
+        public int InvoiceNumber { get; set; }
         public DateTime DueDate { get; set; }
         public string CustomerCountry { get; set; }
         public string OrderId { get; set; }
